// src/app/@core/models/swap.ts
export class Swap {
	swapId: any;
	createdBy = "";
	uploaderPhone = "";
	schoolFrom = "";
	datePosted: any;
	postedBy = "";
	subject1Name: any;
	subject1Id: any;
	subject2Name: any;
	subject2Id: any;
	countyFrom: any;
	subcountyFrom: any;
	countyTo: any;
	subcountyTo: any;
	displayUploaderContact = true;
	countyFromName?: string;    // Only used when viewing
	subCountyFromName?: string; // Only used when viewing
	countyToName?: string;      // Only used when viewing
	subCountyToName?: string;   // Only used when viewing
}